+++ zerver/views/webhooks/__init__.py
@@ -2,9 +2,9 @@
 from __future__ import absolute_import
 from zerver.models import get_client
 from zerver.lib.actions import check_send_message
-from zerver.lib.response import json_success, json_error
+from zerver.lib.response import json_success
 from zerver.decorator import REQ, \
-    has_request_variables, authenticated_rest_api_view, \
+    has_request_variables, \
     api_key_only_webhook_view
 
 import pprint
@@ -13,30 +13,7 @@ import ujson
 from .github import build_commit_list_content, build_message_from_gitlog
 
 
-def truncate(string, length):
-    if len(string) > length:
-        string = string[:length-3] + '...'
-    return string
 
-@authenticated_rest_api_view
-def api_zendesk_webhook(request, user_profile):
-    """
-    Zendesk uses trigers with message templates. This webhook uses the
-    ticket_id and ticket_title to create a subject. And passes with zendesk
-    user's configured message to zulip.
-    """
-    try:
-        ticket_title = request.POST['ticket_title']
-        ticket_id = request.POST['ticket_id']
-        message = request.POST['message']
-        stream = request.POST.get('stream', 'zendesk')
-    except KeyError as e:
-        return json_error('Missing post parameter %s' % (e.message,))
-
-    subject = truncate('#%s: %s' % (ticket_id, ticket_title), 60)
-    check_send_message(user_profile, get_client('ZulipZenDeskWebhook'), 'stream',
-                       [stream], subject, message)
-    return json_success()
 
 
 PAGER_DUTY_EVENT_NAMES = {
