+++ zproject/urls.py
@@ -163,7 +163,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^api/v1/external/desk$',          'webhooks.deskdotcom.api_deskdotcom_webhook'),
     url(r'^api/v1/external/stash$',         'webhooks.stash.api_stash_webhook'),
     url(r'^api/v1/external/freshdesk$',     'webhooks.freshdesk.api_freshdesk_webhook'),
-    url(r'^api/v1/external/zendesk$',       'webhooks.api_zendesk_webhook'),
+    url(r'^api/v1/external/zendesk$',       'webhooks.zendesk.api_zendesk_webhook'),
     url(r'^api/v1/external/pagerduty$',     'webhooks.api_pagerduty_webhook'),
     url(r'^api/v1/external/travis$',        'webhooks.api_travis_webhook'),
